Section: ACP

+++ plugins/newsletter/admin/template/usergroup.php
@@ -43,29 +43,36 @@
             <th></th>
             <th></th>
             <th>
-              <button type="submit" name="delete" id="button_delete" class="btn btn-danger btn-xs"
-                      onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;"><i
-                  class="fa fa-trash-o"></i></button>
+              <button type="submit" name="delete" id="button_delete" class="btn btn-danger btn-xs" onclick="if(!confirm('<?php echo $tl["user"]["al"]; ?>'))return false;">
+                <i class="fa fa-trash-o"></i>
+              </button>
             </th>
           </tr>
           </thead>
           <?php if (isset($JAK_USERGROUP_ALL) && is_array($JAK_USERGROUP_ALL)) foreach ($JAK_USERGROUP_ALL as $v) { ?>
             <tr>
               <td><?php echo $v["id"]; ?></td>
-              <td><input type="checkbox" name="jak_delete_usergroup[]" class="highlight"
-                         value="<?php echo $v["id"]; ?>"/></td>
-              <td><a
-                  href="index.php?p=newsletter&amp;sp=usergroup&amp;ssp=edit&amp;sssp=<?php echo $v["id"]; ?>"><?php echo $v["name"]; ?></a>
+              <td><input type="checkbox" name="jak_delete_usergroup[]" class="highlight" value="<?php echo $v["id"]; ?>"/></td>
+              <td>
+                <a href="index.php?p=newsletter&amp;sp=usergroup&amp;ssp=edit&amp;sssp=<?php echo $v["id"]; ?>"><?php echo $v["name"]; ?></a>
               </td>
               <td><?php echo $v["description"]; ?></td>
-              <td><a href="index.php?p=newsletter&amp;sp=user&amp;ssp=group&amp;sssp=<?php echo $v["id"]; ?>"
-                     class="btn btn-default btn-xs"><i class="fa fa-user"></i></a></td>
-              <td><a href="index.php?p=newsletter&amp;sp=usergroup&amp;ssp=edit&amp;sssp=<?php echo $v["id"]; ?>"
-                     class="btn btn-default btn-xs"><i class="fa fa-edit"></i></a></td>
-              <td><?php if ($v["id"] != 1) { ?><a
-                  href="index.php?p=newsletter&amp;sp=usergroup&amp;ssp=delete&amp;sssp=<?php echo $v["id"]; ?>"
-                  onclick="if(!confirm('<?php echo $tl["user"]["alg"]; ?>'))return false;"
-                  class="btn btn-default btn-xs"><i class="fa fa-trash-o"></i></a><?php } ?></td>
+              <td>
+                <a href="index.php?p=newsletter&amp;sp=user&amp;ssp=group&amp;sssp=<?php echo $v["id"]; ?>" class="btn btn-default btn-xs">
+                  <i class="fa fa-user"></i>
+                </a>
+              </td>
+              <td>
+                <a href="index.php?p=newsletter&amp;sp=usergroup&amp;ssp=edit&amp;sssp=<?php echo $v["id"]; ?>" class="btn btn-default btn-xs">
+                  <i class="fa fa-edit"></i>
+                </a>
+              </td>
+              <td>
+                <?php if ($v["id"] != 1) { ?>
+                  <a href="index.php?p=newsletter&amp;sp=usergroup&amp;ssp=delete&amp;sssp=<?php echo $v["id"]; ?>" data-confirm="<?php echo $tl["user"]["alg"]; ?>" class="btn btn-default btn-xs">
+                    <i class="fa fa-trash-o"></i>
+                  </a>
+                <?php } ?></td>
             </tr>
           <?php } ?>
         </table>
